+++ client/src/test/java/org/ehrbase/client/openehrclient/defaultrestclient/AqlTestIT.java
@@ -44,6 +44,7 @@ import org.ehrbase.client.aql.field.EhrFields;
 import org.ehrbase.client.aql.field.NativeSelectAqlField;
 import org.ehrbase.client.aql.parameter.Parameter;
 import org.ehrbase.client.aql.parameter.ParameterValue;
+import org.ehrbase.client.aql.parameter.StoredQueryParameter;
 import org.ehrbase.client.aql.query.EntityQuery;
 import org.ehrbase.client.aql.query.Query;
 import org.ehrbase.client.aql.record.Record2;
@@ -55,6 +56,7 @@ import org.ehrbase.client.flattener.Flattener;
 import org.ehrbase.client.openehrclient.OpenEhrClient;
 import org.ehrbase.client.templateprovider.TestDataTemplateProvider;
 import org.ehrbase.response.openehr.QueryResponseData;
+import org.ehrbase.response.openehr.StoredQueryResponseData;
 import org.ehrbase.serialisation.jsonencoding.CanonicalJson;
 import org.ehrbase.test_data.composition.CompositionTestDataCanonicalJson;
 import org.junit.After;
@@ -353,7 +355,7 @@ public class AqlTestIT {
 
         List<Record2<UUID, Double>> result =
                 openEhrClient.aqlEndpoint().execute(query, new ParameterValue("ehr_id", ehr));
-        assertThat(result).isNotNull().hasSize(1);
+        assertThat(result).isNotNull().hasSize(2);
     }
 
     @Test
@@ -372,7 +374,7 @@ public class AqlTestIT {
         openEhrClient.compositionEndpoint(ehr).mergeCompositionEntity(coronaAnamneseComposition);
 
         Query<Record2<String, String>> query = Query.buildNativeQuery(
-                "Select DISTINCT"
+                "Select "
                         + " o/data[at0001]/events[at0002]/data[at0003]/items[at0022]/items[at0005]/value/value as var1,"
                         + " o/data[at0001]/events[at0002]/data[at0003]/items[at0022]/items[at0004]/value/value as var2"
                         + " from EHR e[ehr_id/value = $ehr_id] "
@@ -431,7 +433,7 @@ public class AqlTestIT {
 
         List<Record2<UUID, Double>> result =
                 openEhrClient.aqlEndpoint().execute(entityQuery, ehrIdParameter.setValue(ehr));
-        assertThat(result).isNotNull().size().isEqualTo(1);
+        assertThat(result).isNotNull().size().isEqualTo(2);
     }
 
     @Test
@@ -528,5 +530,213 @@ public class AqlTestIT {
         String actualMessage = exception.getMessage();
 
         assertTrue(actualMessage.contains(expectedMessage));
+    }
+
+    @Test
+    public void testExecuteStoredAqlQueryInvalid() {
+        performInvalidExecuteStoredQuery(null);
+        performInvalidExecuteStoredQuery(new StoredQueryParameter(null, null));
+        performInvalidExecuteStoredQuery(new StoredQueryParameter("queryName", null));
+        performInvalidExecuteStoredQuery(new StoredQueryParameter(null, "version"));
+        performInvalidExecuteStoredQuery(new StoredQueryParameter("org.openehr::blablabla", null));
+    }
+
+    private void performInvalidExecuteStoredQuery(StoredQueryParameter requestParam) {
+        Exception exception = assertThrows(
+                ClientException.class, () -> openEhrClient.aqlEndpoint().executeStoredQuery(requestParam));
+
+        String expectedMessage = "Invalid query";
+        String actualMessage = exception.getMessage();
+
+        assertTrue(actualMessage.contains(expectedMessage));
+    }
+
+    @Test
+    public void testStoreAqlQueryInvalid1() {
+        performInvalidStoreAqlQuery(null, null, "Invalid query");
+        performInvalidStoreAqlQuery(null, new StoredQueryParameter(null, null), "Invalid query");
+        performInvalidStoreAqlQuery(null, new StoredQueryParameter("queryName", null), "Invalid query");
+        performInvalidStoreAqlQuery(null, new StoredQueryParameter(null, "version"), "Invalid query");
+
+        Query<Record2<UUID, Double>> query = Query.buildNativeQuery(
+                "select e/ehr_id/value,o/data[at0001]/events[at0002]/data[at0003]/items[at0004]/value/magnitude  "
+                        + "from EHR e[ehr_id/value = $ehr_id]  "
+                        + "contains COMPOSITION a [openEHR-EHR-COMPOSITION.sample_encounter.v1] contains Observation o[openEHR-EHR-OBSERVATION.sample_blood_pressure.v1]"
+                        + "where o/data[at0001]/events[at0002]/data[at0003]/items[at0004]/value/magnitude = 1.1",
+                UUID.class,
+                Double.class);
+
+        performInvalidStoreAqlQuery(query, null, "Invalid parameters");
+        performInvalidStoreAqlQuery(query, new StoredQueryParameter(null, null), "Invalid parameters");
+        performInvalidStoreAqlQuery(query, new StoredQueryParameter("queryName", null), "Invalid parameters");
+        performInvalidStoreAqlQuery(query, new StoredQueryParameter(null, "version"), "Invalid parameters");
+    }
+
+    private void performInvalidStoreAqlQuery(Query query, StoredQueryParameter requestParam, String expectedMessage) {
+        Exception exception = assertThrows(
+                ClientException.class, () -> openEhrClient.aqlEndpoint().storeAqlQuery(query, requestParam));
+
+        assertTrue(exception.getMessage().contains(expectedMessage));
+    }
+
+    @Test
+    public void testGetAqlQueryInvalid() {
+        checkInvalidStoredAqlQuery(null, null);
+        checkInvalidStoredAqlQuery(null, "");
+        checkInvalidStoredAqlQuery("", null);
+        checkInvalidStoredAqlQuery("", "");
+    }
+
+    private void checkInvalidStoredAqlQuery(String qualifiedQueryName, String version) {
+        StoredQueryParameter requestParam = new StoredQueryParameter(qualifiedQueryName, version);
+
+        Exception exception = assertThrows(
+                ClientException.class, () -> openEhrClient.aqlEndpoint().getStoredAqlQuery(requestParam));
+
+        String expectedMessage = "Invalid query";
+
+        assertTrue(exception.getMessage().contains(expectedMessage));
+    }
+
+    @Test
+    public void testStoreAndRetrieveAqlQuery() {
+
+        Query<Record2<UUID, Double>> query = Query.buildNativeQuery(
+                "SELECT c " + "FROM EHR e[ehr_id/value=$ehr_id] "
+                        + "CONTAINS COMPOSITION c[openEHR-EHR-COMPOSITION.encounter.v1] "
+                        + "CONTAINS OBSERVATION obs[openEHR-EHR-OBSERVATION.blood_pressure.v1] "
+                        + "WHERE obs/data[at0001]/events[at0006]/data[at0003]/items[at0004]/value/magnitude >= 100",
+                UUID.class,
+                Double.class);
+
+        String qualifiedQueryName = "org.openehr::storedQueryName";
+        int patchVersion = 0;
+        String version = "1.0.";
+
+        boolean successful = false;
+
+        do {
+            // try to store the query
+            try {
+                openEhrClient
+                        .aqlEndpoint()
+                        .storeAqlQuery(query, new StoredQueryParameter(qualifiedQueryName, version + patchVersion));
+                successful = true;
+            } catch (Exception e) {
+                patchVersion++;
+            }
+        } while (!successful && patchVersion < 1000);
+
+        String fullVersion = version + patchVersion;
+
+        StoredQueryResponseData storedAqlQuery = openEhrClient
+                .aqlEndpoint()
+                .getStoredAqlQuery(new StoredQueryParameter(qualifiedQueryName, fullVersion));
+
+        // TODO: the qualified query name should not have the version number included
+        String expectedName = qualifiedQueryName + "/" + version + patchVersion;
+        assertEquals(expectedName, storedAqlQuery.getName());
+
+        // TODO: version is currently included within the qualified name and the version field is not filled
+        // this seems to not be conform to the specification
+        // https://specifications.openehr.org/releases/ITS-REST/Release-1.0.0/definitions.html#definitions-stored-query-get-1 (for Release 1.0.0)
+        // https://specifications.openehr.org/releases/ITS-REST/Release-1.0.2/definitions.html#definitions-stored-query-get-1 (for currently valid Release 1.0.2)
+        // assertEquals(fullVersion,storedAqlQuery.getVersion());
+        assertEquals(query.buildAql(), storedAqlQuery.getAqlQuery());
+    }
+
+    @Test
+    public void testExecuteStoredAqlQuery() {
+
+        ehr = openEhrClient.ehrEndpoint().createEhr();
+
+        Query<Record2<UUID, Double>> query = Query.buildNativeQuery(
+                "SELECT c " + "FROM EHR e[ehr_id/value=$ehr_id] "
+                        + "CONTAINS COMPOSITION c[openEHR-EHR-COMPOSITION.encounter.v1] "
+                        + "CONTAINS OBSERVATION obs[openEHR-EHR-OBSERVATION.blood_pressure.v1] "
+                        + "WHERE obs/data[at0001]/events[at0006]/data[at0003]/items[at0004]/value/magnitude >= 100",
+                UUID.class,
+                Double.class);
+
+        String qualifiedQueryName = "org.openehr::storedQueryName";
+        int patchVersion = 0;
+        String version = "1.0.";
+
+        boolean successful = false;
+
+        do {
+            // try to store the query
+            try {
+                openEhrClient
+                        .aqlEndpoint()
+                        .storeAqlQuery(query, new StoredQueryParameter(qualifiedQueryName, version + patchVersion));
+                successful = true;
+            } catch (Exception e) {
+                patchVersion++;
+            }
+        } while (!successful && patchVersion < 1000);
+
+        String fullVersion = version + patchVersion;
+
+        QueryResponseData queryResponse = openEhrClient
+                .aqlEndpoint()
+                .executeStoredQuery(new StoredQueryParameter(qualifiedQueryName, fullVersion));
+
+        // TODO: the qualified query name should not have the version number included
+        String expectedName = qualifiedQueryName + "/" + version + patchVersion;
+        assertEquals(expectedName, queryResponse.getName());
+
+        assertEquals(query.buildAql(), queryResponse.getQuery());
+
+        assertTrue(queryResponse.getColumns().size() > 0);
+    }
+
+    @Test
+    public void testExecuteStoredAqlQueryFullParams() {
+
+        ehr = openEhrClient.ehrEndpoint().createEhr();
+
+        Query<Record2<UUID, Double>> query = Query.buildNativeQuery(
+                "SELECT c " + "FROM EHR e[ehr_id/value=$ehr_id] "
+                        + "CONTAINS COMPOSITION c[openEHR-EHR-COMPOSITION.encounter.v1] "
+                        + "CONTAINS OBSERVATION obs[openEHR-EHR-OBSERVATION.blood_pressure.v1] "
+                        + "WHERE obs/data[at0001]/events[at0006]/data[at0003]/items[at0004]/value/magnitude >= 100",
+                UUID.class,
+                Double.class);
+
+        String qualifiedQueryName = "org.openehr::storedQueryName";
+        int patchVersion = 0;
+        String version = "1.0.";
+
+        boolean successful = false;
+
+        do {
+            // try to store the query
+            try {
+                openEhrClient
+                        .aqlEndpoint()
+                        .storeAqlQuery(query, new StoredQueryParameter(qualifiedQueryName, version + patchVersion));
+                successful = true;
+            } catch (Exception e) {
+                patchVersion++;
+            }
+        } while (!successful && patchVersion < 1000);
+
+        String fullVersion = version + patchVersion;
+
+        int offsetValue = 1;
+        int fetchValue = 2;
+        StoredQueryParameter queryParameter = new StoredQueryParameter(qualifiedQueryName, fullVersion);
+        queryParameter.offset(offsetValue).fetch(fetchValue).addQueryParam("ehr_id", ehr.toString());
+
+        QueryResponseData queryResponse = openEhrClient.aqlEndpoint().executeStoredQuery(queryParameter);
+
+        // TODO: the qualified query name should not have the version number included
+        String expectedName = qualifiedQueryName + "/" + version + patchVersion;
+        assertEquals(expectedName, queryResponse.getName());
+
+        assertEquals(query.buildAql() + " LIMIT " + fetchValue + " OFFSET " + offsetValue, queryResponse.getQuery());
+
+        assertTrue(queryResponse.getColumns().size() > 0);
     }
 }