+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 
 ## [unreleased]
  ### Added 
+   - Added handling of stored AQL query requests ([#384](https://github.com/ehrbase/openEHR_SDK/pull/384))
 
  ### Fixed 
 - update archie  to 2.1.0 ([#410](https://github.com/ehrbase/openEHR_SDK/pull/410))